(4b/) Move Ry() to structure.math.Ry()
Keeping things uniform within the samplers module
